Rename <name>.hpp into <KokkosFFT_Distributed_name>.hpp
