--- distributed/src/Distributed.hpp
@@ -1,7 +1,0 @@
-#ifndef DISTRIBUTED_HPP
-#define DISTRIBUTED_HPP
-
-#include "Plan.hpp"
-#include "MPI_Helper.hpp"
-
-#endif
