+++ distributed/src/KokkosFFT_Distributed.hpp
@@ -1,0 +1,7 @@
+#ifndef KOKKOSFFT_DISTRIBUTED_HPP
+#define KOKKOSFFT_DISTRIBUTED_HPP
+
+#include "KokkosFFT_Distributed_Plan.hpp"
+#include "KokkosFFT_Distributed_MPI_Helper.hpp"
+
+#endif
