+++ distributed/src/KokkosFFT_Distributed_CufftMp_Transform.hpp
@@ -1,5 +1,5 @@
-#ifndef CUFFT_MP_TRANSFORM_HPP
-#define CUFFT_MP_TRANSFORM_HPP
+#ifndef KOKKOSFFT_DISTRIBUTED_CUFFT_MP_TRANSFORM_HPP
+#define KOKKOSFFT_DISTRIBUTED_CUFFT_MP_TRANSFORM_HPP
 
 #include <Kokkos_Core.hpp>
 #include <KokkosFFT.hpp>
