+++ distributed/src/KokkosFFT_Distributed_RocfftMPI_Transform.hpp
@@ -1,5 +1,5 @@
-#ifndef ROCFFT_MPI_TRANSFORM_HPP
-#define ROCFFT_MPI_TRANSFORM_HPP
+#ifndef KOKKOSFFT_DISTRIBUTED_ROCFFT_MPI_TRANSFORM_HPP
+#define KOKKOSFFT_DISTRIBUTED_ROCFFT_MPI_TRANSFORM_HPP
 
 #include <Kokkos_Core.hpp>
 #include <KokkosFFT.hpp>
