+++ distributed/src/KokkosFFT_Distributed_SharedPlan.hpp
@@ -1,9 +1,9 @@
-#ifndef SHARED_PLAN_HPP
-#define SHARED_PLAN_HPP
+#ifndef KOKKOSFFT_DISTRIBUTED_SHARED_PLAN_HPP
+#define KOKKOSFFT_DISTRIBUTED_SHARED_PLAN_HPP
 
 #include <Kokkos_Core.hpp>
 #include <KokkosFFT.hpp>
-#include "InternalPlan.hpp"
+#include "KokkosFFT_Distributed_InternalPlan.hpp"
 
 template <typename ExecutionSpace, typename InViewType, typename OutViewType,
           std::size_t DIM = 1, typename InLayoutType = Kokkos::LayoutRight,
