+++ distributed/unit_test/Test_Extents.cpp
@@ -2,7 +2,7 @@
 #include <gtest/gtest.h>
 #include <iostream>
 #include <Kokkos_Core.hpp>
-#include "Extents.hpp"
+#include "KokkosFFT_Distributed_Extents.hpp"
 #include "Test_Utils.hpp"
 
 namespace {
